requirements.txt
streamlit>=1.30.0
firebase-admin>=6.4.0
Pillow>=10.0.0
